+++ Model/Category/Attribute/Source/CategoryMapping.php
@@ -5,6 +5,7 @@ namespace Monext\Payline\Model\Category\Attribute\Source;
 use Magento\Framework\Component\ComponentRegistrar;
 use Magento\Framework\File\Csv;
 use Monext\Payline\Helper\Constants as HelperConstants;
+use Monext\Payline\Helper\Data as PaylineHelper;
 
 /**
  * This class serve to map Magento Categories to Payline Categories : https://docs.payline.com/display/DT/Codes+-+Category
@@ -20,13 +21,16 @@ class CategoryMapping extends \Magento\Eav\Model\Entity\Attribute\Source\Abstrac
      * @var \Magento\Framework\File\Csv
      */
     protected $csvReader;
+    private $helperPayline;
 
     public function __construct(
         ComponentRegistrar $componentRegistrar,
-        Csv $csvReader
+        Csv $csvReader,
+        PaylineHelper $helperPayline
     ) {
         $this->componentRegistrar = $componentRegistrar;
         $this->csvReader = $csvReader;
+        $this->helperPayline = $helperPayline;
     }
 
     /**
@@ -37,12 +41,9 @@ class CategoryMapping extends \Magento\Eav\Model\Entity\Attribute\Source\Abstrac
     {
         if (!$this->_options) {
             $this->_options = [];
-
-            $csvFile = $this->componentRegistrar->getPath(ComponentRegistrar::MODULE, HelperConstants::MODULE_NAME) . '/fixtures/category_mapping.csv';
-            $rows = $this->csvReader->getData($csvFile);
-
+            $rows = $this->helperPayline->getDefaultCategories();
             foreach ($rows as $row) {
-                $this->_options[] = ['value' => $row[0], 'label' => __($row[1])];
+                $this->_options[] = ['value' => $row['value'], 'label' => __($row['name'])];
             }
 
             array_unshift($this->_options, ['value' => '', 'label' => __('Please select a category mapping...')]);